Added ability to suppress SPEC file generate
This seems like an odd flag to add to this tool.
However, sometimes you want to use this tool to gen a SPEC (--spec-only), make some changes to the SPEC by hand (patch notes, etc...), and then generate the RPM.
This will let you use whatever SPEC exists in the build root.